fix MyAdapter - demo plotter
